+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 - Simplify API for flushing events ([#2030](https://github.com/getsentry/sentry-dotnet/pull/2030))
 
+### Fixes
+
+- Update unobserved task exception integration ([#2034](https://github.com/getsentry/sentry-dotnet/pull/2034))
+- Fix trace propagation targets setter ([#2035](https://github.com/getsentry/sentry-dotnet/pull/2035))
+
 ## 3.23.1
 
 ### Fixes